+++ zio-http-benchmarks/src/main/scala/zhttp.benchmarks/HttpRouteTextPerf.scala
@@ -15,7 +15,7 @@ class HttpRouteTextPerf {
 
   private val res          = Response.text("HELLO WORLD")
   private val app          = Http.succeed(res)
-  private val req: Request = Request(Method.GET, URL6(!!))
+  private val req: Request = Request(Method.GET, URL(!!))
   private val httpProgram  = ZIO.foreach_(0 to 1000) { _ => app.execute(req).toZIO }
   private val UIOProgram   = ZIO.foreach_(0 to 1000) { _ => UIO(res) }
