+++ zio-http/src/main/scala/zhttp/http/URL.scala
@@ -1,77 +1,134 @@
 package zhttp.http
 
 import io.netty.handler.codec.http.{QueryStringDecoder, QueryStringEncoder}
-import zhttp.http.URL.Fragment
+import zhttp.http.Scheme.HTTP
+import zhttp.http.URL._
 
 import java.net.URI
 import scala.jdk.CollectionConverters._
 import scala.util.Try
 
-final case class URL(
-  path: Path,
-  kind: URL.Location = URL.Location.Relative,
-  queryParams: Map[String, List[String]] = Map.empty,
-  fragment: Option[Fragment] = None,
-) { self =>
-  val host: Option[String] = kind match {
-    case URL.Location.Relative      => None
-    case abs: URL.Location.Absolute => Option(abs.host)
+sealed trait URL { self =>
+
+  def getHost: Option[String]   = self match {
+    case b: Absolute => b.host
+    case r           => r.toAbsolute.getHost
+  }
+  def getPort: Option[Int]      = self match {
+    case b: Absolute => b.port
+    case r           => r.toAbsolute.getPort
+  }
+  def getScheme: Option[Scheme] = self match {
+    case b: Absolute => b.scheme
+    case r           => r.toAbsolute.getScheme
   }
 
-  val port: Option[Int] = kind match {
-    case URL.Location.Relative      => None
-    case abs: URL.Location.Absolute => Option(abs.port)
+  def toAbsolute: Absolute = self match {
+    case Unsafe(x)   => URL.unsafeFromString(x)
+    case b: Absolute => b
+    case c: Relative => Absolute(relative = c)
   }
 
-  private[zhttp] def relative: URL = self.kind match {
-    case URL.Location.Relative => self
-    case _                     => self.copy(kind = URL.Location.Relative)
-  }
-
-  def encode: String = URL.asString(self)
-
-  def setPath(path: Path) =
-    copy(path = path)
-
-  def setPath(path: String) = copy(path = Path(path))
-
-  def setHost(host: String) = {
-    val location = kind match {
-      case URL.Location.Relative      => URL.Location.Absolute(Scheme.HTTP, host, URL.portFromScheme(Scheme.HTTP))
-      case abs: URL.Location.Absolute => abs.copy(host = host)
-    }
-    copy(kind = location)
-  }
-
-  def setQueryParams(queryParams: Map[String, List[String]]) =
-    copy(queryParams = queryParams)
-
-  def setQueryParams(query: String) =
-    copy(queryParams = URL.queryParams(query))
-
-  def setPort(port: Int) = {
-    val location = kind match {
-      case URL.Location.Relative      => URL.Location.Absolute(Scheme.HTTP, "", port)
-      case abs: URL.Location.Absolute => abs.copy(port = port)
-    }
-
-    copy(kind = location)
-  }
-
-  def setScheme(scheme: Scheme) = {
-    val location = kind match {
-      case URL.Location.Relative      => URL.Location.Absolute(scheme, "", URL.portFromScheme(scheme))
-      case abs: URL.Location.Absolute => abs.copy(scheme = scheme)
-    }
-
-    copy(kind = location)
-  }
+  def setHost(host: String): URL     =
+    self.toAbsolute.copy(Some(host), Some(self.getScheme.getOrElse(HTTP)), Some(self.getPort.getOrElse(80)))
+  def setPort(port: Int): URL        =
+    self.toAbsolute.copy(Some(self.getHost.getOrElse("localhost")), Some(self.getScheme.getOrElse(HTTP)), Some(port))
+  def setScheme(scheme: Scheme): URL =
+    self.toAbsolute.copy(Some(self.getHost.getOrElse("localhost")), Some(scheme), Some(self.getPort.getOrElse(80)))
+  def setPath(path: Path): URL       = self.toAbsolute.copy(relative = self.toAbsolute.relative.copy(path = path))
+  def setPath(path: String): URL     = self.toAbsolute.copy(relative = self.toAbsolute.relative.copy(path = Path(path)))
+  def setQueryParams(queryParams: Map[String, List[String]]): URL =
+    self.toAbsolute.copy(relative = self.toAbsolute.relative.copy(queryParams = queryParams))
+  def setQueryParams(query: String): URL                          =
+    self.toAbsolute.copy(relative = self.toAbsolute.relative.copy(queryParams = URL.queryParams(query)))
+  def encode: String                                              = URL.asString(self)
 }
 object URL {
-  sealed trait Location
-  object Location {
-    case object Relative                                               extends Location
-    final case class Absolute(scheme: Scheme, host: String, port: Int) extends Location
+
+  def apply(path: Path): URL     = Relative(path)
+  def apply(string: String): URL = Unsafe(string)
+
+  def asString(url: URL): String        = {
+    val p: String = path(url.toAbsolute.relative)
+    url match {
+      case u: Absolute =>
+        if (u.scheme.isDefined && u.port.isDefined && u.host.isDefined) {
+          if (u.port.get == 80 || u.port.get == 443)
+            s"${u.scheme.get.encode}://${u.host.get}$p"
+          else s"${u.scheme.get.encode}://${u.host.get}:${u.port.get}$p"
+        } else p
+      case _           => {
+        p
+      }
+    }
+  }
+  private def path(r: Relative): String = {
+    val encoder = new QueryStringEncoder(s"${r.path.encode}${r.fragment.fold("")(f => "#" + f.raw)}")
+    r.queryParams.foreach { case (key, values) =>
+      if (key != "") values.foreach { value => encoder.addParam(key, value) }
+    }
+    encoder.toString
+  }
+
+  def fromString(string: String): Either[HttpError.BadRequest, URL] = Try(unsafeFromString(string)).toEither match {
+    case Right(value) if value != null => Right(value)
+    case _                             => Left(HttpError.BadRequest(s"Invalid URL: $string"))
+  }
+
+  final case class Unsafe(string: String) extends URL
+  final case class Absolute(
+    host: Option[String] = None,
+    scheme: Option[Scheme] = None,
+    port: Option[Int] = None,
+    relative: Relative = Relative(),
+  ) extends URL
+
+  final case class Relative(
+    path: Path = !!,
+    queryParams: Map[String, List[String]] = Map.empty,
+    fragment: Option[Fragment] = None,
+  ) extends URL
+
+  def unsafeFromString(string: String): Absolute =
+    try {
+      val url = new URI(string)
+      if (url.isAbsolute) unsafeFromAbsoluteURI(url) else unsafeFromRelativeURI(url)
+    } catch {
+      case _: Throwable => throw new Error("Invalid URL")
+    }
+
+  private def unsafeFromAbsoluteURI(uri: URI): Absolute = {
+
+    def portFromScheme(scheme: Scheme): Int = scheme match {
+      case Scheme.HTTP  => 80
+      case Scheme.HTTPS => 443
+      case null         => -1
+    }
+
+    val scheme  = Scheme.fromString2(uri.getScheme)
+    val uriPort = uri.getPort
+    val port    = if (uriPort != -1) uriPort else portFromScheme(scheme)
+    val host    = uri.getHost
+
+    if (port != -1 && scheme != null && host != null)
+      Absolute(
+        Some(host),
+        Some(scheme),
+        Some(port),
+        Relative(Path(uri.getRawPath), queryParams(uri.getRawQuery), Fragment.fromURI(uri)),
+      )
+    else throw new Error()
+  }
+
+  private def unsafeFromRelativeURI(uri: URI): Absolute =
+    Absolute(relative = Relative(Path(uri.getRawPath), queryParams(uri.getRawQuery), Fragment.fromURI(uri)))
+
+  case class Fragment private (raw: String, decoded: String)
+  object Fragment {
+    def fromURI(uri: URI): Option[Fragment] = for {
+      raw     <- Option(uri.getRawFragment)
+      decoded <- Option(uri.getFragment)
+    } yield Fragment(raw, decoded)
   }
 
   private def queryParams(query: String) = {
@@ -84,67 +141,16 @@ object URL {
     }
   }
 
-  private def portFromScheme(scheme: Scheme): Int = scheme match {
-    case Scheme.HTTP  => 80
-    case Scheme.HTTPS => 443
-  }
-
-  private def fromAbsoluteURI(uri: URI): Option[URL] = {
-    for {
-      scheme <- Scheme.fromString(uri.getScheme)
-      host   <- Option(uri.getHost)
-      path   <- Option(uri.getRawPath)
-      port       = Option(uri.getPort).filter(_ != -1).getOrElse(portFromScheme(scheme))
-      connection = URL.Location.Absolute(scheme, host, port)
-    } yield URL(Path(path), connection, queryParams(uri.getRawQuery), Fragment.fromURI(uri))
-  }
-
-  private def fromRelativeURI(uri: URI): Option[URL] = for {
-    path <- Option(uri.getRawPath)
-  } yield URL(Path(path), Location.Relative, queryParams(uri.getRawQuery), Fragment.fromURI(uri))
-
+  def root: URL  = Absolute(relative = Relative())
   def empty: URL = root
 
-  def fromString(string: String): Either[HttpError, URL] = {
-    def invalidURL = Left(HttpError.BadRequest(s"Invalid URL: $string"))
-    for {
-      url <- Try(new URI(string)).toEither match {
-        case Left(_)      => invalidURL
-        case Right(value) => Right(value)
-      }
-      url <- (if (url.isAbsolute) fromAbsoluteURI(url) else fromRelativeURI(url)) match {
-        case None        => invalidURL
-        case Some(value) => Right(value)
-      }
+  val url1 = URL.empty
+    .setHost("www.zio-http.com")
+    .setQueryParams(Map("A" -> List("B")))
+    .setPort(8090)
+    .setScheme(Scheme.HTTP)
 
-    } yield url
-  }
+  val url2: URL                               = URL("www.zio-http.com/a")
+  val url3: Either[HttpError.BadRequest, URL] = URL.fromString("www.zio-http.com/a")
 
-  def asString(url: URL): String = {
-
-    def path: String = {
-      val encoder = new QueryStringEncoder(s"${url.path.encode}${url.fragment.fold("")(f => "#" + f.raw)}")
-      url.queryParams.foreach { case (key, values) =>
-        if (key != "") values.foreach { value => encoder.addParam(key, value) }
-      }
-      encoder.toString
-    }
-
-    url.kind match {
-      case Location.Relative                     => path
-      case Location.Absolute(scheme, host, port) =>
-        if (port == 80 || port == 443) s"${scheme.encode}://$host$path"
-        else s"${scheme.encode}://$host:$port$path"
-    }
-  }
-
-  case class Fragment private (raw: String, decoded: String)
-  object Fragment {
-    def fromURI(uri: URI): Option[Fragment] = for {
-      raw     <- Option(uri.getRawFragment)
-      decoded <- Option(uri.getFragment)
-    } yield Fragment(raw, decoded)
-  }
-
-  def root: URL = URL(!!)
 }
