+++ zio-http/src/test/scala/zhttp/http/middleware/CorsSpec.scala
@@ -13,7 +13,7 @@ object CorsSpec extends DefaultRunnableSpec with HttpAppTestExtensions {
     testM("OPTIONS request") {
       val request = Request(
         method = Method.OPTIONS,
-        url = URL6(!! / "success"),
+        url = URL(!! / "success"),
         headers = Headers.accessControlRequestMethod(Method.GET) ++ Headers.origin("test-env"),
       )
 
@@ -35,7 +35,7 @@ object CorsSpec extends DefaultRunnableSpec with HttpAppTestExtensions {
         val request =
           Request(
             method = Method.GET,
-            url = URL6(!! / "success"),
+            url = URL(!! / "success"),
             headers = Headers.accessControlRequestMethod(Method.GET) ++ Headers.origin("test-env"),
           )
